feat: change image in catalog on hover
