+++ src/types/types.tsx
@@ -1,5 +1,7 @@
+import { Image } from '@commercetools/platform-sdk';
+
 export type CategoryCardProps = {
-  images?: string;
+  images?: Image[];
   keyCard: string;
   prices?: number | string;
   discounted?: number | string;
